strutturazione progetto
- creato subfolders per modulo src diviso per sito su cui fare scraping
- sistemato il main program
